Split test_proposal_api to separate test cases
